Remove React Component from posts reducer to common/Message.tsx
